pyproject.toml
[project]
name = "coremind"
version = "0.0.1"
description = "A framework for continuous personal intelligence."
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
authors = [{ name = "Guillaume Gagnon" }]
requires-python = ">=3.12"
keywords = ["ai", "agent", "cognitive", "personal-intelligence"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "aiohttp>=3.10.0",
    "aioimaplib>=1.1.0",
    "click>=8.1.0",
    "cryptography>=43.0.0",
    "grpcio>=1.66.0",
    "jcs>=0.2.1",
    "jinja2>=3.1.0",
    "litellm>=1.50.0",
    "pydantic>=2.9.0",
    "qdrant-client>=1.11.0",
    "structlog>=24.4.0",
    "surrealdb>=1.0.0",
]

[project.scripts]
coremind = "coremind.cli:cli"

[project.optional-dependencies]
embeddings-local = ["sentence-transformers>=3.0.0"]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "jsonschema>=4.23.0",
    "grpcio>=1.66.0",
    "grpcio-tools>=1.66.0",
    "grpc-stubs>=1.53.0",
    "types-protobuf>=5.0.0",
    "pydantic>=2.9.0",
    "psutil>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/gagnongui/coremind"
Repository = "https://github.com/gagnongui/coremind"
Documentation = "https://github.com/gagnongui/coremind/tree/main/docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coremind"]

# --------------------------------------------------------------------
# Ruff
# --------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]
extend-exclude = ["**/_generated/**"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "N",     # pep8-naming
    "SIM",   # flake8-simplify
    "PL",    # pylint
    "RUF",   # ruff-specific
    "ANN",   # flake8-annotations
    "ASYNC",
    "S",     # bandit
    "A",     # flake8-builtins
    "COM",   # flake8-commas
    "DTZ",   # flake8-datetimez — enforce timezone-aware datetimes
    "PIE",
    "RET",
    "TID",
    "PTH",   # flake8-use-pathlib
    "ERA",   # eradicate
    "T20",   # flake8-print
]
ignore = [
    "COM812",  # conflicts with formatter
    "ANN401",  # allow `Any` in rare places with justification
    "PLR0913", # too many arguments — sometimes necessary
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",
    "PLR2004",
    "ANN",
] # assert/magic-values/annotations are fine in tests
"scripts/**/*.py" = [
    "T20",
    "S603",
    "S607",
] # CLI scripts are allowed to print and shell out

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --------------------------------------------------------------------
# Mypy
# --------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
show_error_codes = true
pretty = true
exclude = ["_generated"]
mypy_path = [
    "src",
    "plugins/systemstats",
    "plugins/homeassistant",
    "plugins/gmail-imap",
    "integrations/openclaw-adapter/coremind_side",
]

[[tool.mypy.overrides]]
module = [
    "coremind.plugin_api._generated.*",
    "coremind_plugin_openclaw._generated.*",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["sentence_transformers", "aioimaplib", "litellm"]
ignore_missing_imports = true

# --------------------------------------------------------------------
# Pytest
# --------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = [
    "src",
    "plugins/systemstats",
    "plugins/homeassistant",
    "plugins/gmail-imap",
    "integrations/openclaw-adapter/coremind_side",
]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src/coremind",
    "--cov-report=term-missing",
]
markers = [
    "integration: tests requiring external services (docker-compose)",
    "e2e: full end-to-end scenarios",
    "slow: tests that take > 1 second",
    "unit: fast unit tests with no I/O",
]

[tool.coverage.run]
branch = true
source = ["src/coremind"]
omit = ["*/_generated/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
